@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace ImagoOpus\Actions;
 
@@ -6,11 +6,25 @@ use Imagick;
 use ImagoOpus\Image;
 use ImagickPixel;
 
-class Matrix extends AAction
+class Matrix implements ActionInterface
 {
-    public function run(Image $image)
+    /** @var array<int, float>  */
+    private array $matrix;
+    private string $background;
+
+    /**
+     * @param array<int, float> $matrix
+     * @param string $background Background color if image don't support alpha
+     */
+    public function __construct(array $matrix, string $background = '#fff')
     {
-        $matrix = $this->options['matrix'];
+        $this->matrix = $matrix;
+        $this->background = $background;
+    }
+
+    public function run(Image $image): Image
+    {
+        $matrix = $this->matrix;
 
         $angle = atan2($matrix[1], $matrix[0]) * (180 / M_PI);
 
@@ -19,7 +33,7 @@ class Matrix extends AAction
 
         $skew = atan2($matrix[3], $matrix[2]) * (180 / M_PI);
         $skew = round($skew - $angle);
-        if ($skew==-90) {
+        if ($skew == -90) {
             $image->flipImage();
         }
 
@@ -27,13 +41,13 @@ class Matrix extends AAction
         $tPixel = new ImagickPixel('transparent');
         do {
             $image->rotateImage($tPixel, $angle);
-        } while ($image->nextImage());
+        }
+        while ($image->nextImage());
 
         if (!$image->hasAlpha()) {
-            $backgroundColor = $this->options['background']?:'#fff';
             $dim = $image->getImageGeometry();
             $background = new Imagick();
-            $background->newImage($dim['width'], $dim['height'], new ImagickPixel($backgroundColor));
+            $background->newImage($dim['width'], $dim['height'], new ImagickPixel($this->background));
             $currentFormat = $image->getImageFormat();
             $image->setImageFormat("png");
             $image->compositeImage($background, Imagick::COMPOSITE_DSTATOP, 0, 0);
