@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace ImagoOpus\Actions;
 
@@ -6,38 +6,31 @@ use ImagoOpus\Image;
 use ImagickPixel;
 use Imagick;
 
-class Rotate extends AAction
+/**
+ * Action to free rotate image
+ *
+ * Default is to auto rotate image from image meta data
+ */
+class Rotate implements ActionInterface
 {
-    protected function autoRotate(Image $image)
+    private ?float $angle;
+    private string $background;
+
+    /**
+     * @param float|null $angle If angle is null it will be auto rotated to ORIENTATION_TOPLEFT
+     * @param string $background Background color if image don't support alpha
+     */
+    public function __construct(float $angle = null, string $background = '#fff')
     {
-        $orientation = $image->getImageOrientation();
-
-        switch ($orientation) {
-            case Imagick::ORIENTATION_BOTTOMRIGHT:
-                $image->rotateimage("#000", 180); // rotate 180 degrees
-                break;
-
-            case Imagick::ORIENTATION_RIGHTTOP:
-                $image->rotateimage("#000", 90); // rotate 90 degrees CW
-                break;
-
-            case Imagick::ORIENTATION_LEFTBOTTOM:
-                $image->rotateimage("#000", -90); // rotate 90 degrees CCW
-                break;
-        }
-
-        $image->setImageOrientation(Imagick::ORIENTATION_TOPLEFT);
-
-        return $image;
+        $this->angle = $angle;
+        $this->background = $background;
     }
 
-    public function run(Image $image)
+    public function run(Image $image): Image
     {
-        if (!isset($this->options['angle'])) {
+        if (!$this->angle) {
             return $this->autoRotate($image);
         }
-        $angle = $this->options['angle'];
-
         $currentFormat = $image->getImageFormat();
         if (!$image->hasAlpha()) {
             $image->setImageFormat("png");
@@ -45,17 +38,43 @@ class Rotate extends AAction
         $image->setIteratorIndex(0);
         $tPixel = new ImagickPixel('transparent');
         do {
-            $image->rotateimage($tPixel, $angle);
-        } while ($image->nextImage());
-        
+            $image->rotateImage($tPixel, $this->angle);
+        }
+        while ($image->nextImage());
+
         if (!$image->hasAlpha()) {
-            $backgroundColor = $this->options['background']?:'#fff';
             $dim = $image->getImageGeometry();
             $background = new Imagick();
-            $background->newImage($dim['width'], $dim['height'], new ImagickPixel($backgroundColor));
+            $background->newImage($dim['width'], $dim['height'], new ImagickPixel($this->background));
             $image->compositeImage($background, Imagick::COMPOSITE_DSTATOP, 0, 0);
             $image->setImageFormat($currentFormat);
         }
+
+        return $image;
+    }
+
+    protected function autoRotate(Image $image): Image
+    {
+        $orientation = $image->getImageOrientation();
+
+        switch ($orientation) {
+            case Imagick::ORIENTATION_BOTTOMRIGHT:
+                // rotate 180 degrees
+                $image->rotateImage("#000", 180);
+                break;
+
+            case Imagick::ORIENTATION_RIGHTTOP:
+                // rotate 90 degrees CW
+                $image->rotateImage("#000", 90);
+                break;
+
+            case Imagick::ORIENTATION_LEFTBOTTOM:
+                // rotate 90 degrees CCW
+                $image->rotateImage("#000", -90);
+                break;
+        }
+
+        $image->setImageOrientation(Imagick::ORIENTATION_TOPLEFT);
 
         return $image;
     }
