@@ -1,20 +1,29 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace ImagoOpus\Effects;
 
 use ImagoOpus\Image;
-use ImagoOpus\Actions\AEffect;
 
-class Blur extends AEffect
+class Blur implements EffectInterface
 {
-    public function run(Image $image)
+    use FixRangeTrait;
+
+    private float $radius;
+
+    public function __construct(float $radius)
     {
-        $radius = $this->_fixRange($this->options['radius'], 0, 100) / 10;
-        
+        $this->radius = $radius;
+    }
+
+    public function run(Image $image): Image
+    {
+        $radius = $this->fixRange($this->radius, 0, 100) / 10;
+
         $image->setIteratorIndex(0);
         do {
             $image->blurImage($radius, 100);
-        } while ($image->nextImage());
+        }
+        while ($image->nextImage());
 
         return $image;
     }
