@@ -1,25 +1,37 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace ImagoOpus\Effects;
 
+use Imagick;
 use ImagoOpus\Image;
-use ImagoOpus\Actions\AEffect;
 
-class Contrast extends AEffect
+class Contrast implements EffectInterface
 {
-    public function run(Image $image)
+    use FixRangeTrait;
+
+    private float $contrast;
+
+    public function __construct(float $contrast)
     {
-        $contrast = $this->_fixRange($this->options['contrast'], -100, 100);
-        
+        $this->contrast = $contrast;
+    }
+
+    public function run(Image $image): Image
+    {
+        $contrast = $this->fixRange($this->contrast, -100, 100);
+
         $image->setIteratorIndex(0);
-        if (method_exists('Imagick', 'brightnessContrastImage')) {
+        if (method_exists(Imagick::class, 'brightnessContrastImage')) {
             do {
                 $image->brightnessContrastImage(0, $contrast);
-            } while ($image->nextImage());
-        } else {
+            }
+            while ($image->nextImage());
+        }
+        else {
             do {
-                $image->sigmoidalContrastImage($contrast > 0 ? true : false, abs($contrast), 0.5);
-            } while ($image->nextImage());
+                $image->sigmoidalContrastImage($contrast > 0, abs($contrast), 0.5);
+            }
+            while ($image->nextImage());
         }
 
         return $image;
