@@ -1,21 +1,30 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace ImagoOpus\Effects;
 
 use ImagoOpus\Image;
-use ImagoOpus\Actions\AEffect;
 
-class Sharpen extends AEffect
+class Sharpen implements EffectInterface
 {
-    public function run(Image $image)
+    use FixRangeTrait;
+
+    private float $amount;
+
+    public function __construct(float $amount = 5)
     {
-        $amount = $this->_fixRange($this->options['amount'] ?: 5, 5, 1000);
+        $this->amount = $amount;
+    }
+
+    public function run(Image $image): Image
+    {
+        $amount = $this->fixRange($this->amount, 5, 1000);
         $amount = ($amount * 3) / 100;
 
         $image->setIteratorIndex(0);
         do {
             $image->sharpenimage(0, $amount);
-        } while ($image->nextImage());
+        }
+        while ($image->nextImage());
 
         return $image;
     }
