@@ -1,23 +1,36 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace ImagoOpus\Effects;
 
 use ImagoOpus\Image;
-use ImagoOpus\Actions\AEffect;
 
-class Vignette extends AEffect
+class Vignette implements EffectInterface
 {
-    public function run(Image $image)
+    use FixRangeTrait;
+
+    private float $blur;
+    private int $x;
+    private int $y;
+
+    public function __construct(float $blur = 10, int $x = 10, int $y = 10)
+    {
+        $this->blur = $blur;
+        $this->x = $x;
+        $this->y = $y;
+    }
+
+    public function run(Image $image): Image
     {
         $sigma = 10;
-        $blur = $this->_fixRange($this->options['blur'] ?: 10, 0, 255);
-        $x = $this->_fixRange($this->options['x'] ?: 10, 0, 1000);
-        $y = $this->_fixRange($this->options['y'] ?: 10, 0, 1000);
+        $blur = $this->fixRange($this->blur, 0, 255);
+        $x = $this->fixRange($this->x, 0, 1000);
+        $y = $this->fixRange($this->y, 0, 1000);
 
         $image->setIteratorIndex(0);
         do {
-            $image->vignetteImage($blur, $sigma, $x, $y);
-        } while ($image->nextImage());
+            $image->vignetteImage($blur, $sigma, (int)$x, (int)$y);
+        }
+        while ($image->nextImage());
 
         return $image;
     }
